requirements.txt
asgiref==3.2.3
cacheout==0.11.2
cachetools==4.0.0
Django==3.0.1
djangorestframework==3.11.0
ttlru-dict==1.0.1
